==> 02-data.Rmd <==
# Data sources

We collected our data from **Statista**, a German company specializing in market and consumer data. The specific dataset is from its digital market outlook - digital advertising section.

The link is provided here:
[link](https://www.statista.com/outlook/dmo/digital-advertising/worldwide)

We selected United States, United Kingdom and China as our observation units and respectively downloaded the dataset.

Datasets are present as following:
```{r, echo = FALSE, message = FALSE, warning = FALSE}
library(tidyverse)
china.orginal <- read_csv("dmo_digital-advertising_usd_china_en_2022-03-03_18-16-31.csv")
us.original <- read_csv("dmo_digital-advertising_usd_united states_en_2022-03-03_18-15-03.csv")
uk.original <- read_csv("dmo_digital-advertising_usd_united kingdom_en_2022-03-03_18-10-30.csv")
head(china.orginal)
head(us.original)
head(us.original)
```

Each dataset has 16 columns: Region, Market, Chart, Name, Unit, Source and Years (2017-2026). The first five columns are filled with character data, and Years columns are filled with values which are all numeric data.

**Region**

It is a column filled with region. Each dataset has one region, so it does not help much when we research in one country. But it will be helpful when we combine datasets to do the comparison. 

**Market**

```{r, echo = FALSE, message = FALSE, warning = FALSE}
unique(china.orginal$Market)
```

There are six factors inside this column, with the whole market "Digital Advertising", and five markets inside it: "Banner", "Social Media", "Search", "Classifieds", and "Video". 

**Chart**

```{r, echo = FALSE, message = FALSE, warning = FALSE}
unique(china.orginal$Chart)
```

This column contains different charts we would use for further analysis. Mainly, we would look at Ad Spending, Users, and Reach by Social Network charts.

**Name**

This column contains different Names of Charts. We will not present them here because the dataset has not been transformed to tidy version.

**Unit, Source, Year**

Those columns are obvious, so we will not talk about them in detail here.

***

Briefly, we noticed that this dataset is not in tidy format and have several NAs. We will clean the dataset in the next sections.
