==> 05-results.Rmd <==
# Results

In this section, we will analysis the digital advertising market and provide our findings. 
```{r echo=FALSE, message=FALSE, warning=FALSE}
library(ggplot2)
library(ggthemes)
library(GGally)
library(CGPfunctions)
library(tidyverse)
library(ggbump)
library(reshape2)
library(gt)
library(reactable)
library(reactablefmtr)
library(scales)
library(waffle)
library(treemapify)
library(highcharter)
library(rCharts)
library(knitr)
us <- read.csv("us.csv")
uk <- read.csv("uk.csv")
china <- read.csv("china.csv")
df <- rbind(us,uk,china)
```

### Research Question 1:

Firstly, we look at the Average Ad spending and how it is divided into desktop and mobile. 

```{r echo=FALSE, message=FALSE, warning=FALSE}
library(webr)
library(tidyverse)
dfx<-df[df$Chart=="Ad Spending"&df$Name!="Total"&df$Market!="Classifieds",]
dfx<-dplyr::select(dfx,Region,Market,Name,Year,Value)
dfx$Name <- str_replace_all(dfx$Name,c("Banner Advertising Desktop"="Desktop","Banner Advertising Mobile"="Mobile"))
dfx$Name <- str_replace_all(dfx$Name,c("Social Media Advertising Desktop"="Desktop","Social Media Advertising Mobile"="Mobile"))
dfx$Name <- str_replace_all(dfx$Name,c("Search Advertising Desktop"="Desktop","Search Advertising Mobile"="Mobile"))
dfx$Name <- str_replace_all(dfx$Name,c("Video Advertising Desktop"="Desktop","Video Advertising Mobile"="Mobile"))
dfx$Market <- str_replace_all(dfx$Market,"Banner Advertising","Banner")
dfx$Market <- str_replace_all(dfx$Market,"Social Media Advertising","Social")
dfx$Market <- str_replace_all(dfx$Market,"Search Advertising","Search")
dfx$Market <- str_replace_all(dfx$Market,"Video Advertising","Video")
dfw <- dfx%>%group_by(Market,Name,Year)%>%
  summarize(sum(Value))
colnames(dfw)[4]<-"Total"
dfp<-dfw%>%group_by(Market,Name)%>%
  mutate(mean_p=mean(Total[Year%in%c(2017,2018,2019,2020,2021)]))%>%
  filter(Year==2017)%>%
  select(-c(Year,Total))
PieDonut(dfp,aes(Market,Name,count=mean_p),title="Average Ad Spending for the Past (2017-2021)")

dfy<-dfw%>%group_by(Market,Name)%>%
  mutate(mean_f=mean(Total[Year%in%c(2022,2023,2024,2025,2026)]))%>%
  filter(Year==2022)%>%
  select(-c(Year,Total))
PieDonut(dfy,aes(Market,Name,count=mean_f),title="Average Ad Spending for the Future (2022-2026)")
```

Since Classifieds(including General, Jobs, Motor, Real Estate) represents only a small portion of the overall Ad spending across all digital advertising marketplace, we only include the four major markets: Video, Banner, Social Media, and Search Advertising. We separately plot the donut charts for the past(2017-2021) and the future(predicted value for 2022-2026). For both of them, we calculated the average ad spending over 5 years and present it as the percentage in donut charts.

Search and Banner Advertising market currently dominates the digital market; In terms of forecasts, Social Media and Video Advertising spending will increase and each four major markets will become more even across the whole digital market, while Search Advertising market remains the leading spender and Video Advertising market is the lowest. 

By the comparison between the donut chart of the past and the future, it is clear that the mobile section of Ad spending is increasing in all four markets. In particular, in Video Advertising, mobile will increase its share of AD spending by nearly 20%. Mobile AD spending in the banner and search AD markets will also increase by more than 10%. Social media advertising had the lowest growth rate, at about 5 percent, which makes sense given that mobile advertising already accounts for more than 80 percent of AD spending. 

```{r echo=FALSE, message=FALSE, warning=FALSE}
dfa <- dfx%>%group_by(Region,Market,Name,Year)%>%
  summarize(sum(Value))
colnames(dfa)[5]<-"Total"
dfa$Region<-factor(dfa$Region,levels=c("United Kingdom","China","United States"),labels=c("UK","China","US"))

dfb<-dfa%>%group_by(Region,Market,Name)%>%
  mutate(mean_p=mean(Total[Year%in%c(2017,2018,2019,2020,2021)]))%>%
  filter(Year==2017)%>%
  select(-c(Year,Total))

library(ggalluvial)
ggplot(dfb, aes(y = mean_p, axis1 = fct_relevel(Region,"China",after=1), axis2 = Market, axis3=Name)) +
  geom_alluvium(aes(fill = Name), width = 1/12) +
  geom_stratum(width = 1/12, fill = "grey80", color = "black") +
  geom_label(stat = "stratum", aes(label = after_stat(stratum))) +
  ggtitle("Average Ad Spending for the Past(2017-2021)")+
  labs(y="Ad Spending  in million USD")+
  scale_x_discrete(limit = c("Country", "Market","M/D"), expand = c(.05, .05)) +
  scale_y_continuous(expand = c(.01,0)) +
  guides(fill = FALSE) +
  theme_classic()


dfc<-dfa%>%group_by(Region,Market,Name)%>%
  mutate(mean_p=mean(Total[Year%in%c(2022,2023,2024,2025,2026)]))%>%
  filter(Year==2022)%>%
  select(-c(Year,Total))

ggplot(dfc, aes(y = mean_p, axis1 = fct_relevel(Region,"China",after=1), axis2 = Market, axis3=Name)) +
  geom_alluvium(aes(fill = Name), width = 1/12) +
  geom_stratum(width = 1/12, fill = "grey80", color = "black") +
  geom_label(stat = "stratum", aes(label = after_stat(stratum))) +
  ggtitle("Average Ad Spending for the Future(2022-2026)")+
  labs(y="Ad Spending  in million USD")+
  scale_x_discrete(limit = c("Country", "Market","M/D"), expand = c(.05, .05)) +
  scale_y_continuous(expand = c(.01,0)) +
  guides(fill = FALSE) +
  theme_classic()
```

The alluvial diagrams are also separately plotted by the timeline(the past and the future). The alluviums are colored by Desktop and Mobile. The difference between the alluvial diagrams and donut charts is that we can take a closer look at the change in the proportion of mobile and desktop sections in different countries. As we can see, the overall proportion of the mobile section will increase a lot from the past to the future. From countries' sides, all three countries will increase their mobile AD spending in all four markets. More specifically, by seeing the change in width of alluviums of the mobile section, UK will increase its mobile Ad spending most in Video Advertising market; China will increase its mobile Ad spending most in Social Media Advertising market; US will increase its mobile Ad spending most in Video Advertising market. 

In conclusion, the weight of ad spending will gradually shift towards the mobile side, which is in line with our expectations. We believe that with the rapid development of mobile Internet, more and more Internet users can be exposed to mobile advertising, so mobile advertising will dominate digital advertising. 

### Research Question 2:

Now, we will analysis the ad spending in different countries and in different markets. 

```{r echo=FALSE, message=FALSE, warning=FALSE}
df5 <- df[df$Chart=="Ad Spending"&df$Name=="Total",]
ggplot(df5,aes(x=Year, y=Value,fill=fct_reorder(Region,Value,.desc=FALSE)))+
  geom_col(position="dodge")+
  facet_wrap(~Market,scale="free",nrow=5,ncol=1)+
  ggtitle("Ad Spending in different markets (2017-2026) between three countries")+
  labs(fill="Country",y="Ad Spending in million USD")
```

Firstly, we'd like to get a general idea of how much advertising is spent in different markets in the three countries. As you can see from the boxplots, overall Ad spending is increasing in all three countries, regardless of markets. However, the differences between the three countries vary in five markets. For Classifields, China has consistently been the highest spender of the three countries and will continue to grow at a faster rate than the US and UK in terms of projected value (2022-2026). In Search and Video Advertising market, the United States has been far ahead of the other two countries in terms of Ad spending, twice as much as the next largest country, China. In Social Media Advertising market, the US is also the leader, with China following at an even faster pace. In addition, The United Kingdom has consistently lagged far behind the other countries in five markets. 


```{r echo=FALSE, message=FALSE, warning=FALSE}
Tspending <- df %>% 
  filter(df$Market=="Social Media Advertising"|df$Market=="Video Advertising"|df$Market=="Search Advertising")%>%
  filter(Chart=="Ad Spending"&Name=="Total")
Tspending <- dplyr::select(Tspending,2:4,Year,Value)
df2 <- df %>% 
  filter(df$Market=="Social Media Advertising"|df$Market=="Video Advertising"|df$Market=="Search Advertising")%>%
  filter(Chart=="Top Company Revenues (Worldwide & Consolidated)")
Trevenue <- aggregate(df2$Value,list(df2$Market,df2$Chart,df2$Year), FUN=sum, na.rm=TRUE)
colnames(Trevenue) <- (c("Market","Chart","Year","Value"))
Trevenue <- Trevenue[Trevenue$Year%in%c(2017,2018,2019,2020), ]
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
library(ggthemes)

theme_heat <- theme_classic() +
  theme(axis.line = element_blank(),
        axis.ticks = element_blank())

Tspending <- Tspending %>% group_by(Market, Chart, Year)%>%
  summarise(sum(Value))
 colnames(Tspending)[4]<-"Value"
df3 <- rbind(Tspending[Tspending$Year%in%c(2017,2018,2019,2020),],Trevenue) 

df4 <- df3%>%group_by(Market,Year)%>%
  mutate(Return=0.5*Value[Chart=="Top Company Revenues (Worldwide & Consolidated)"]/Value[Chart=="Ad Spending"])
df4 <- df4[1:12,]
ggplot(df4, aes(Year,Market))+
  geom_tile(aes(fill=Return),color="white")+
  coord_fixed()+
  scale_fill_viridis_c()+
  theme_heat

```

We calculated the return on investment ratio (revenue divided by spending) based on available data on AD spending versus revenue(only available in Video, Social Media, and Search Advertising markets) from 2017 to 2020. Since returns are globally aggregated, we assume that these three countries account for 50% of worldwide revenue (the accuracy of this ratio does not affect the observation of change in return ratio). 

According to the heatmap, we find that the return ratio of Video and Social Media Advertising decreases year by year, and the rate of return of social media decreases faster given the more pronounced degree of color change for Social Media Advertising, while the return ratio of Search Advertising increases year by year. This finding is so interesting that we further take a closer look at advertising spending growth rate and spending per internet user for each market over the years.

```{r echo=FALSE, message=FALSE, warning=FALSE}
library(highcharter)
df6 <- df[df$Chart=="Ad Spending Growth"&df$Name=="Total",]


df6i <- df6%>%select(-c(X,Chart,Name,Unit))%>%pivot_wider(names_from = Market, values_from = Value)%>%
  filter(Year!=2017)

```

```{r echo=FALSE, message=FALSE, warning=FALSE}
china.total.1721 <- china %>% filter(Chart == "Average Ad Spending per Internet User" & Name == "Total" & Year == c(2017:2021)) %>% select(c(Region, Market, Year, Value))
china.total.1721.wide <- china.total.1721 %>% pivot_wider(names_from = Year, values_from = Value)

us.total.1721 <- us %>% filter(Chart == "Average Ad Spending per Internet User" & Name == "Total" & Year == c(2017:2021)) %>% select(c(Region, Market, Year, Value))
us.total.1721.wide <- us.total.1721 %>% pivot_wider(names_from = Year, values_from = Value)

uk.total.1721 <- uk %>% filter(Chart == "Average Ad Spending per Internet User" & Name == "Total" & Year == c(2017:2021)) %>% select(c(Region, Market, Year, Value))
uk.total.1721.wide <- uk.total.1721 %>% pivot_wider(names_from = Year, values_from = Value)

total.1721.wide <- rbind(china.total.1721.wide,us.total.1721.wide,uk.total.1721.wide)
total.1721.wide$Region <- factor(total.1721.wide$Region, levels = c("China", "United States", "United Kingdom"))

total.1721 <- rbind(china.total.1721, us.total.1721, uk.total.1721)
```


```{r echo=FALSE, message=FALSE, warning=FALSE}
china.total.2226 <- china %>% filter(Chart == "Average Ad Spending per Internet User" & Name == "Total" & Year == c(2022:2026)) %>% select(c(Region, Market, Year, Value))
china.total.2226.wide <- china.total.2226 %>% pivot_wider(names_from = Year, values_from = Value)

us.total.2226 <- us %>% filter(Chart == "Average Ad Spending per Internet User" & Name == "Total" & Year == c(2022:2026)) %>% select(c(Region, Market, Year, Value))
us.total.2226.wide <- us.total.2226 %>% pivot_wider(names_from = Year, values_from = Value)

uk.total.2226 <- uk %>% filter(Chart == "Average Ad Spending per Internet User" & Name == "Total" & Year == c(2022:2026)) %>% select(c(Region, Market, Year, Value))
uk.total.2226.wide <- uk.total.2226 %>% pivot_wider(names_from = Year, values_from = Value)

total.2226.wide <- rbind(china.total.2226.wide,us.total.2226.wide,uk.total.2226.wide)
total.2226.wide$Region <- factor(total.2226.wide$Region, levels = c("China", "United States", "United Kingdom"))

total.2226 <- rbind(china.total.2226, us.total.2226, uk.total.2226)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
total.1721.avg <- total.1721 %>% group_by(Region, Market) %>% summarize(Average = mean(Value)) %>% mutate(Year = "2017-2021")

total.2226.avg <- total.2226 %>% group_by(Region, Market) %>% summarize(Average = mean(Value)) %>% mutate(Year = "2022-2026")

total.avg <- rbind(total.1721.avg, total.2226.avg)%>% arrange(Region, Market)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
total.avg.china <- total.avg %>% filter(Region=="China") %>% select(Year, Market, Average)
newggslopegraph(dataframe = total.avg.china, Times = Year, Measurement = Average, Grouping = Market, Title = "China Ad Spending Per Internet User Evolution", SubTitle = "Present vs. Projected",Caption = NULL)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
c <- highchart() %>% 
  hc_xAxis(categories = df6i[df6i$Region=="China", ]$Year) %>% 
  hc_add_series(name = "Banner", 
                data = df6i[df6i$Region=="China", ]$`Banner Advertising`) %>% 
  hc_add_series(name = "Social Media", 
                data = df6i[df6i$Region=="China", ]$`Social Media Advertising`) %>%
  hc_add_series(name = "Search", 
                data = df6i[df6i$Region=="China", ]$`Search Advertising`) %>%
  hc_add_series(name = "Classifieds", 
                data = df6i[df6i$Region=="China", ]$Classifieds) %>%
  hc_add_series(name = "Video", 
                data = df6i[df6i$Region=="China", ]$`Video Advertising`) 
c <- c %>%
  hc_title(text = "Ad Spending Growth Rate for Markets in China",
           margin = 20, 
           align = "left",
           style = list(color = "steelblue")) %>% 
  hc_subtitle(text = "2018 to 2026",
              align = "left",
              style = list(color = "#2b908f", 
                           fontWeight = "bold")) %>% 
  hc_credits(enabled = TRUE, # add credits
             text = "Gapminder Data",
             href = "http://gapminder.com") %>% 
  hc_legend(align = "left", 
            verticalAlign = "top",
            layout = "vertical", 
            x = 0, 
            y = 100) %>%
  hc_tooltip(crosshairs = TRUE, 
             backgroundColor = "#FCFFC5",
             shared = TRUE, 
             borderWidth = 4) %>% 
  hc_exporting(enabled = TRUE)

c
```

(If you move your mouse over the points in the plot, you will see the growth rate for each market in one year)

China: \
The overall trend of Ad Spending per Internet User across all market is increasing by projection. Social Media Advertising has the largest projected growth rate, while search advertising does not have a large projected growth rate. Though, search advertising has the largest ad spending per internet user for now, it seems that market will focus on Social Media Advertising and Banner Advertising in the future.\
The overall trend for Ad spending growth across all markets is a year-over-year decline, with a plateau starting in 2023. Ultimately, Video and Social Media Advertising will have similar and higher growth rates than that of another three advertising markets, and the other three markets will have almost similar growth rates. In addition, we also notice that the ad spending growth Video, Classifields and Search Advertising have an obvious increase from 2020 to 2021, among which search advertising has the highest proportion of increase.

```{r echo=FALSE, message=FALSE, warning=FALSE}
total.avg.us <- total.avg %>% filter(Region=="United States") %>% select(Year, Market, Average)
total.avg.us
newggslopegraph(dataframe = total.avg.us, Times = Year, Measurement = Average, Grouping = Market, Title = "US Ad Spending Per Internet User Evolution", SubTitle = "Present vs. Projected",Caption = NULL)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
s <- highchart() %>% 
  hc_xAxis(categories = df6i[df6i$Region=="United States", ]$Year) %>% 
  hc_add_series(name = "Banner", 
                data = df6i[df6i$Region=="United States", ]$`Banner Advertising`) %>% 
  hc_add_series(name = "Social Media", 
                data = df6i[df6i$Region=="United States", ]$`Social Media Advertising`) %>%
  hc_add_series(name = "Search", 
                data = df6i[df6i$Region=="United States", ]$`Search Advertising`) %>%
  hc_add_series(name = "Classifieds", 
                data = df6i[df6i$Region=="United States", ]$Classifieds) %>%
  hc_add_series(name = "Video", 
                data = df6i[df6i$Region=="United States", ]$`Video Advertising`) 
s <- s %>%
  hc_title(text = "Ad Spending Growth Rate for Markets in United States",
           margin = 20, 
           align = "left",
           style = list(color = "steelblue")) %>% 
  hc_subtitle(text = "2018 to 2026",
              align = "left",
              style = list(color = "#2b908f", 
                           fontWeight = "bold")) %>% 
  hc_credits(enabled = TRUE, # add credits
             text = "Gapminder Data",
             href = "http://gapminder.com") %>% 
  hc_legend(align = "left", 
            verticalAlign = "top",
            layout = "vertical", 
            x = 0, 
            y = 100) %>%
  hc_tooltip(crosshairs = TRUE, 
             backgroundColor = "#FCFFC5",
             shared = TRUE, 
             borderWidth = 4) %>% 
  hc_exporting(enabled = TRUE)

s
```

US: \
The overall trend of Ad Spending per Internet User across all market is increasing by projection except for Classified. Other Advertising has similar projected growth rate, but Banner Advertising has a smaller projected growth rate on Ad Spending. It is clear that US market is still focusing on Search Advertising.\
The overall trend for Ad spending growth across all markets is a year-over-year decline, with a flattening out starting in 2024. Finally, the growth rates of Video, Search, and Social Media Advertising are similar and higher than that of the other two, followed by Banner and then Classified, which will maintain the lowest growth rate. In addition, Video, Classifields and Search have an obvious increase from 2020 to 2021, among which search has the highest proportion of increase.

```{r echo=FALSE, message=FALSE, warning=FALSE}
total.avg.uk <- total.avg %>% filter(Region=="United Kingdom") %>% select(Year, Market, Average)
total.avg.uk
newggslopegraph(dataframe = total.avg.uk, Times = Year, Measurement = Average, Grouping = Market, Title = "UK Ad Spending Per Internet User Evolution", SubTitle = "Present vs. Projected",Caption = NULL)
```

```{r echo=FALSE, message=FALSE, warning=FALSE}
k <- highchart() %>% 
  hc_xAxis(categories = df6i[df6i$Region=="United Kingdom", ]$Year) %>% 
  hc_add_series(name = "Banner", 
                data = df6i[df6i$Region=="United Kingdom", ]$`Banner Advertising`) %>% 
  hc_add_series(name = "Social Media", 
                data = df6i[df6i$Region=="United Kingdom", ]$`Social Media Advertising`) %>%
  hc_add_series(name = "Search", 
                data = df6i[df6i$Region=="United Kingdom", ]$`Search Advertising`) %>%
  hc_add_series(name = "Classifieds", 
                data = df6i[df6i$Region=="United Kingdom", ]$Classifieds) %>%
  hc_add_series(name = "Video", 
                data = df6i[df6i$Region=="United Kingdom", ]$`Video Advertising`) 
k <- k %>%
  hc_title(text = "Ad Spending Growth Rate for Markets in United Kingdom",
           margin = 20, 
           align = "left",
           style = list(color = "steelblue")) %>% 
  hc_subtitle(text = "2018 to 2026",
              align = "left",
              style = list(color = "#2b908f", 
                           fontWeight = "bold")) %>% 
  hc_credits(enabled = TRUE, # add credits
             text = "Gapminder Data",
             href = "http://gapminder.com") %>% 
  hc_legend(align = "left", 
            verticalAlign = "top",
            layout = "vertical", 
            x = 0, 
            y = 100) %>%
  hc_tooltip(crosshairs = TRUE, 
             backgroundColor = "#FCFFC5",
             shared = TRUE, 
             borderWidth = 4) %>% 
  hc_exporting(enabled = TRUE)

k
```

UK:\
The overall trend of Ad Spending per Internet User across all market is increasing by projection. Search and Social Media Advertising has larger projected growth rate. It is clear that UK market is still focusing on Search Advertising.\
The overall trend for Ad spending growth across all markets is a year-over-year decline, with a plateau starting in 2023. Finally, the growth rates of Video, Search, and Social Media Advertising are similar and higher than that of the other two, followed by Banner and then Classified, which will maintain the lowest growth rate. In addition, we also find that all markets have a significant increase from 2020 to 2021, among which search has the highest proportion of increase.

**Comparison between the above three line plots for each country:**\
1. All three countries experienced rapid declines in Ad spending growth for all markets in 2018-2020, followed by a rebound in some or all markets in 2021, with the UK showing the largest increase.\
2. From the prediction of the future growth, growths in all five markets stabilize around 2023 for all three countries. Among them, Video, Search and Social Media Advertising in developed countries (the United Kingdom and the United States) are stable at 7-9%. However, only Video and Social Media Advertising maintain an increased rate of around 10% in China, a developing country. We speculate the reason behind it is that China has a large proportion of the population and advertising is still in the stage of development, so the growth rate could be higher. In addition, the low spending growth of Search Advertising might be due to the lack of a relatively perfect Search engine in China and the limitation of usage of Google in the country.\
3. Based on the heatmap for the return ratio, the sharp and continuous declination of the return ratio for Social Media Advertising can possibly be a reason which explains the growth rate of social media advertising spending is decreasing year by year for all three countries except for the minimal increase from 2019 to 2021 in the UK.\
4. Based on the heatmap for the return ratio, only the return ratio of Search Advertising kept increasing from 2017 to 2020. The increased return ratio might indicate the reason why Search Advertising has the highest increase in Ad spending growth in 2021 among all three countries.\
5. Based on the slope graph, the US and the UK have a similar situation. They spend mostly on Search Advertising per Internet user, and then Social Media and Banner Advertising per Internet user. China, on the other hand, spends a lot on Social Media advertising and is decreasing its spending on Search Advertising per Internet user.\
6. The US and the UK is expected to see the largest growth in Search Advertising spending per Internet user, which can be well proved by the consistently rise in return ratio for Search Advertising market. Still, China will increase most in the Social Media Advertising spending per Internet user instead of Search Advertising. The reasons behind it might be the low development and limitation of search engines as we mentioned before. We also understand that the return ratio is calculated using global advertising revenue, which might not well represent each country's situation and we'd like to further explore the difference between China and other two countries in social media market. 

### Social Media Advertising

As an interactive digital channel that facilitate the creation and sharing of information, ideas, and interests, social media had over 3.6 billion users in 2020 worldwide. Obviously, digital advertising will focus on social media market. China, US and UK have spent a total of 110,062.6 million USD on Social Media Advertising 2021 alone. And in those countries, this spending is continuously increasing. 

```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
china <- read_csv("china.csv") %>% select(c(2:8))
uk <- read_csv("uk.csv") %>% select(c(2:8))
us <- read_csv("us.csv") %>% select(c(2:8))

china.users <- china %>% filter(Market == "Social Media Advertising", Chart %in% c("Users by Age", "Users by Gender", "Users by Income"), Year == 2021) %>% select(1,3,4,7)

us.users <- us %>% filter(Market == "Social Media Advertising", Chart %in% c("Users by Age", "Users by Gender", "Users by Income"), Year == 2021) %>% select(1,3,4,7)

uk.users <- uk %>% filter(Market == "Social Media Advertising", Chart %in% c("Users by Age", "Users by Gender", "Users by Income"), Year == 2021) %>% select(1,3,4,7)

users <- rbind(china.users, us.users, uk.users) %>% pivot_wider(names_from = Region, values_from = Value)
```

We first look at the demographics of social media. We could see a culture difference in it. In three countries, the largest part of users are in 25-34 years. However, in US and UK, smallest part of users are in 18-24 years, but in China, it is 55-64 years. Also, in China, there are more male users but in UK and US shows the opposite. When we check by Income levels, it is more interesting that China has more medium income users, US has more low income users, and UK has more high income users.
```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
users %>%
  gt() %>%
  tab_header(
    title = md("**Demographics**"),
    subtitle = "unit in %"
  ) %>%
  cols_align(
  align ="left",
  columns = c(1,2)
) %>%
  tab_style(
    style = list(
      cell_borders(
        sides = c("bottom"),
        color = "black",
        weight = px(1)
      ),
      cell_borders(
        sides = c("bottom"),
        color = "black",
        weight = px(1)
      )),
    locations = list(
      cells_body(
        columns = 1:5,
        rows = 5
      ),
      cells_body(
        columns = 1:5,
        rows = 7
      )
    )
    )

```


Next, we take a brief look at the Ad spending in Social Media market by Year and Country. Clearly, three countries increase their spending year by year. Compared to US and China, UK spend relatively smaller amoun of money in Ad spending in Social Media Market. 

```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
ad.spending <- rbind(china,us,uk) %>% filter(Market == "Social Media Advertising", Chart == "Ad Spending", Name == "Total") %>% select(1,6,7)
ggplot(ad.spending, aes(x = Year, y = Value, fill = Region)) + 
  geom_bar(stat = "identity") +
  labs(title = "Social Media Advertising", y = "Ad Spending (million USD)")
```

We will dig deeper in the next section.
