==> 06-interactive.Rmd <==
# Interactive component



```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
library(tidyverse)
library(ggbump)
library(reshape2)
library(gt)
library(reactable)
library(reactablefmtr)
library(scales)
library(waffle)
library(treemapify)
library(highcharter)
library(knitr)
```


```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
china <- read_csv("china.csv") %>% select(c(2:8))
uk <- read_csv("uk.csv") %>% select(c(2:8))
us <- read_csv("us.csv") %>% select(c(2:8))
```



```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
china.users <- china %>% filter(Market == "Social Media Advertising", Chart %in% c("Users by Age", "Users by Gender", "Users by Income"), Year == 2021) %>% select(1,3,4,7)

us.users <- us %>% filter(Market == "Social Media Advertising", Chart %in% c("Users by Age", "Users by Gender", "Users by Income"), Year == 2021) %>% select(1,3,4,7)

uk.users <- uk %>% filter(Market == "Social Media Advertising", Chart %in% c("Users by Age", "Users by Gender", "Users by Income"), Year == 2021) %>% select(1,3,4,7)

users <- rbind(china.users, us.users, uk.users) %>% pivot_wider(names_from = Region, values_from = Value)
```


```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE, results = 'asis', comment = NA}
ad.spending <- rbind(china,us,uk) %>% filter(Market == "Social Media Advertising", Chart == "Ad Spending", Name == "Total") %>% select(1,6,7)
```

In order to understand the benefits of Social Media Advertising, it is important to understand the mechanics of social media networks. We decide to look into the social network reach in China, the US and the UK in this section. Social network reach refers to the number of users who have come across a particular content on a social platform.

```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
china.reach <- china %>% 
  filter(Chart == "Reach by social network") %>% 
  select(1,4,6,7) %>% 
  filter(Name != "Total")

china.reach.avg <- china.reach %>% 
  group_by(Name) %>% 
  summarize(`Average Users` = mean(Value)) %>% 
  arrange(desc(`Average Users`)) %>% 
  mutate(perc = `Average Users`/sum(`Average Users`)) %>% 
  add_row(Name = "Total", 
          `Average Users` = sum(.$`Average Users`), 
          perc = sum(.$perc), 
          .before = 1) %>%
  mutate(Percentage = percent(perc, accuracy = 0.001)) %>% 
  select(-3)
```

```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
us.reach <- us %>% 
  filter(Chart == "Reach by social network") %>% 
  select(1,4,6,7) %>% 
  filter(Name != "Total")

us.reach.avg <- us.reach %>% 
  group_by(Name) %>% 
  summarize(`Average Users` = mean(Value)) %>% 
  arrange(desc(`Average Users`)) %>%
  mutate(perc = `Average Users`/sum(`Average Users`)) %>% 
  add_row(Name = "Total", 
          `Average Users` = sum(.$`Average Users`), 
          perc = sum(.$perc), 
          .before = 1) %>% 
  mutate(Percentage = percent(perc, accuracy = 0.001)) %>% 
  select(-3) %>% 
  add_row(Name = c(NA,NA,NA), 
          `Average Users` = c(NA,NA,NA))
```


```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
uk.reach <- uk %>% 
  filter(Chart == "Reach by social network") %>% 
  select(1,4,6,7) %>% 
  filter(Name != "Total")

uk.reach.avg <- uk.reach %>% 
  group_by(Name) %>% 
  summarize(`Average Users` = mean(Value)) %>% 
  arrange(desc(`Average Users`)) %>% 
  mutate(perc = `Average Users`/sum(`Average Users`)) %>% 
  add_row(Name = "Total", 
          `Average Users` = sum(.$`Average Users`), 
          perc = sum(.$perc), 
          .before = 1) %>% 
  mutate(Percentage = percent(perc, accuracy = 0.001)) %>% 
  select(-3) %>% 
  add_row(Name = c(NA,NA,NA,NA), 
          `Average Users` = c(NA,NA,NA,NA))
```


```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
reach.avg <- cbind(china.reach.avg, us.reach.avg, uk.reach.avg)
colnames(reach.avg) <- c("Social Network in China", "Average Users in China","Percentage in China", "Social Network in US", "Average Users in US","Percentage in US","Social Network in UK", "Average Users in UK","Percentage in UK")
```

```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
my_theme <- function(data) {
  tab_options(
    data = data,
    heading.align = "middle",
    heading.title.font.size = 20,
    table.border.top.color = "black",
    column_labels.border.bottom.color = "black",
    column_labels.border.bottom.width= px(2),
    column_labels.font.size = 18
  )
}


reach.avg %>%
  gt() %>%
  tab_spanner(label = "China", columns = 1:3) %>%
  tab_spanner(label = "US", columns = 4:6) %>%
  tab_spanner(label = "UK", columns = 7:9) %>%
  cols_label(
    `Social Network in China` = "Social Network",
    `Average Users in China` = "Average Users",
    `Percentage in China` = "Percentage",
    `Social Network in US` = "Social Network",
    `Average Users in US` = "Average Users",
    `Percentage in US` = "Percentage",
    `Social Network in UK` = "Social Network",
    `Average Users in UK` = "Average Users",
    `Percentage in UK` = "Percentage"
  ) %>% 
  tab_header(
    title = md("**Reach By Network**"),
    subtitle = "unit in million users"
  ) %>%
tab_style(
    style = list(
      cell_fill(color = "grey", alpha = 0.2),
      cell_borders(
        side = c("bottom"), 
        color = "black",
        weight = px(1)
        ),
      cell_text(style="italic")
      ),
    locations = cells_body(
      rows = 1
      )
  ) %>% 
  cols_align(
  align ="left",
  columns = everything()
) %>%
  my_theme()
```

From the table, we could see that China has the largest number of social media users and the UK has the least. The huge difference could be led by the population difference. "Total" in the above table is the sum of all social network users, which includes repeated calculation. Thus, we decide to use the distinct counting of total reach (from the dataset) and calculate the user ratio: 
$Average\space Users = \frac{\sum_{year=2017}^{year=2026} Total\space Users\space }{10}$ \
$User\space Ratio = \frac{Average\space Users}{Population}$ 

|Region  |Average Users |User Ratio
|:------ |:------       |:------
|China   |1016.2646     |0.706
|US      |295.7239      |0.893
|UK      |59.801        |0.881

Now it is clear that the US has the largest number of users and China have the least when taking population into account.

From the Reach by Network table, the popular social media platforms vary from country to country. Especially in China, the top 5 social networks are different from those in the US and the UK. We believe that the reason behind this is the blockage of several social media in China. The only 2 social networks not banned in the US and the UK top 5 are LinkedIn and TikTok. The top 5 social networks in China are all Chinese companies.

Another thing worth noting is that, though people in different countries are using different social media, the ratio of each social network user and total users are similar for the same ranking. For example, the ratio of the most popular social network is around 25% in the three countries.


```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
china.reach.avg.top7 <- china.reach.avg %>% 
  select(1,2) %>% 
  add_row(Name = "Other", 
          `Average Users` = sum(.$`Average Users`[9:19])) %>% 
  slice(-(9:19)) %>% 
  slice(-1)
```
```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
h1 <- china.reach.avg.top7 %>% 
  hchart(
    "treemap", 
    hcaes(x = Name, value = `Average Users`, color = `Average Users`)
    ) %>%
  hc_title(
    text = "Reach by Social Network in China (Top 7)",
    margin = 10,
    align = "left",
    style = list(color = "#6da9df", useHTML = TRUE)
  )
h1
```



```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
china.reach.order <- china.reach %>% 
  group_by(Name) %>% 
  mutate(mean = mean(Value)) %>% 
  arrange(desc(mean), Year) %>% 
  select(-mean) %>% ungroup() %>% 
  mutate(Order = rep(c(1:18), each = 10))

df.china.reach.order <- china.reach.order %>% 
  dplyr::rename(`Social Network` = Name, 
                `Million Users (2017-2026)` = Value) %>% 
  group_by(`Social Network`, Order) %>% 
  summarize(`Million Users (2017-2026)` = list(`Million Users (2017-2026)`)) %>% 
  ungroup() %>% 
  arrange(Order) %>% 
  select(1,3)
```


```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
r1 <- reactable(
  df.china.reach.order,
  columns = list(
    `Social Network` = colDef(maxWidth = 200),
    `Million Users (2017-2026)` = colDef(
      cell = react_sparkline(df.china.reach.order,height = 50,
        tooltip_type = 2,
        show_area = TRUE,
        decimals=2,
        statline = "mean",
        statline_color = "orange",
        statline_label_size = "0.5em")
    )
  )
)
```

```{r echo = FALSE}
r1
```

(If you move your mouse over the points in the plot, you will see the corresponding value)

Now we will dig into each country. In China, the most frequent social network is WeChat from Tencent which is the only instant messaging app in the top 7. We can see that the growth rate of Wechat users is stable. Qzone, on the other hand, has a continuously negative growth rate. The users are decreasing since 2019 and are projected to be continuously decreasing in the future. Qzone is also the product of Tencent and is earlier than WeChat. Unlike WeChat, Qzone is a pure social networking platform where users could write blogs on it. However, this feature alone is not attractive enough to keep users. Another worth noting point is that LinkedIn is projected to have a dramatic decrease in users from 2021 to 2022. LinkedIn China announced that it would shut down the localized version of LinkedIn and launch a new InJobs App focusing on helping China-based professionals find jobs in China in 2021. And this announcement and action lead to a dramatic decrease in users.


```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
us.reach.avg.top7 <- us.reach.avg %>% 
  select(1,2) %>% 
  add_row(Name = "Other", 
          `Average Users` = sum(.$`Average Users`[9:16])) %>% 
  slice(-(9:19)) %>% 
  slice(-1)


h2 <- us.reach.avg.top7 %>% 
  hchart(
    "treemap", 
    hcaes(x = Name, value = `Average Users`, color = `Average Users`)
    )%>%
  hc_title(
    text = "Reach by Social Network in US (Top 7)",
    margin = 10,
    align = "left",
    style = list(color = "#6da9df", useHTML = TRUE)
  )
```

```{r echo = FALSE}
h2
```

```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
us.reach.order <- us.reach %>% 
  group_by(Name) %>% 
  mutate(mean = mean(Value)) %>% 
  arrange(desc(mean), Year) %>% 
  select(-mean) %>% 
  ungroup() %>% 
  mutate(Order = rep(c(1:15), each = 10))

df.us.reach.order <- us.reach.order %>% 
  dplyr::rename(`Social Network` = Name, 
                `Million Users (2017-2026)` = Value) %>% 
  group_by(`Social Network`, Order) %>% 
  summarize(`Million Users (2017-2026)` = list(`Million Users (2017-2026)`)) %>% 
  ungroup() %>% 
  arrange(Order) %>% 
  select(1,3)

r2 <- reactable(
  df.us.reach.order,
  columns = list(
    `Social Network` = colDef(maxWidth = 200),
    `Million Users (2017-2026)` = colDef(
      cell = react_sparkline(df.us.reach.order,height = 50,
        tooltip_type = 2,
        show_area = TRUE,
        decimals=2)
    )
  )
)
```

```{r echo=FALSE}
r2
```

```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
uk.reach.avg.top7 <- uk.reach.avg %>% 
  select(1,2) %>% 
  add_row(Name = "Other", 
          `Average Users` = sum(.$`Average Users`[9:15])) %>% 
  slice(-(9:19)) %>% 
  slice(-1)

h3 <- uk.reach.avg.top7 %>% 
  hchart(
    "treemap", 
    hcaes(x = Name, value = `Average Users`, color = `Average Users`)
    )%>%
  hc_title(
    text = "Reach by Social Network in UK (Top 7)",
    margin = 10,
    align = "left",
    style = list(color = "#6da9df", useHTML = TRUE)
  )
```

```{r echo = FALSE}
h3
```

```{r echo = FALSE, message = FALSE, warning = FALSE, error = FALSE}
uk.reach.order <- uk.reach %>% 
  group_by(Name) %>% 
  mutate(mean = mean(Value)) %>% 
  arrange(desc(mean), Year) %>% 
  select(-mean) %>% 
  ungroup() %>% 
  mutate(Order = rep(c(1:14), each = 10))

df.uk.reach.order <- uk.reach.order %>% 
  dplyr::rename(`Social Network` = Name, 
                `Million Users (2017-2026)` = Value) %>% 
  group_by(`Social Network`, Order) %>% 
  summarize(`Million Users (2017-2026)` = list(`Million Users (2017-2026)`)) %>% 
  ungroup() %>% 
  arrange(Order) %>% 
  select(1,3)


r3 <- reactable(
  df.uk.reach.order,
  columns = list(
    `Social Network` = colDef(maxWidth = 200),
    `Million Users (2017-2026)` = colDef(
      cell = react_sparkline(df.uk.reach.order,height = 50,
        tooltip_type = 2,
        show_area = TRUE,
        decimals=2)
    )
  )
)
```

```{r echo=FALSE}
r3
```

The UK and the US have a similar situation on the social network. Facebook has the most users and keeps a stable growth rate. Pinterest, though ranking six in both countries, has a decreasing growth rate in the the US but a stable growth rate in the UK. Twitter has had a very unstable growth rate in recent years in the US. But the data provider Statista has a positive view of Twitter's user growth rate in the future. The user growth rate of Tumblr is worth noting. Similar to Qzone in China, Tumblr faced a massive drop in users from 2018 to 2019 and continues to drop till now in both countries. In December of 2018, Tumblr announced a new ban on “adult content", and this action resulted in a dramatic drop in users.

**Conclusion**\
Billions of people around the world use social media to share information and make connections, and thus the social media advertising market is becoming stronger. The Social Network market in China is very different from the US and the UK. And it might lead to the reason why China is spending mostly on social media advertising but the US and the UK are not.
